// src/models/rootTraitName.ts
/** Titles of main traits. */
export enum RootTraitName {
  Eyes = 'Eyes',
  Hair = 'Hair',
  Body = 'Body',
  Clothes = 'Clothes',
  Items = 'Items',
  Personality = 'Personality',
  Role = 'Role',
  EngagesIn = 'Engages in',
  SubjectOf = 'Subject of',
  EngagesInSexual = 'Engages in (Sexual)',
  SubjectOfSexual = 'Subject of (Sexual)',
}